# By Konrad Zemek (17) and others
# Via Michał Wrzeszcz (14) and others
* commit 'd53922325d5b2aec6e114bd765f0f9d787d273e9 [formerly a847cca451b6454bfda88b880318d949332168c0]': (40 commits)
  submodules update
  Protocol updated
  Release notes created
  VFS-741 Stop logging certificate verification error when disabled.
  indent change
  change ConnectionPoolInfo args order
  change CommunicationHandler args order
  fix reorder warning
  final fix
  compilation fix
  proto update
  VFS-711 Add default constructor for StorageHelperFactory.
  VFS-711 Add default constructor for StorageHelperFactory.
  VFS-711 Include condition_variable header where needed.
  VFS-711 Include condition_variable header where needed.
  VFS-736 Fix sending log messages to the server.
  VFS-727 Use a persistent boost::asio endpoint.
  update protocol to release
  test fix
  sending mode on file creation
  ...


Former-commit-id: ef1632ded4c73704c449ebe83a525348cca5247e

+++ src/fileCache.cc
@@ -64,7 +64,7 @@ void FileCache::debugPrint()
     multiset<block_ptr>::iterator it = m_fileBlocks.begin();
     while(it != m_fileBlocks.end())
     {
-        printf("\t Offset: %ld, Size: %lu, Data: %s, valid_to: %lu\n", (*it)->offset, (*it)->size(), (*it)->data.c_str(), (*it)->valid_to);
+        printf("\t Offset: %lld, Size: %lu, Data: %s, valid_to: %llu\n", (long long int) (*it)->offset, (*it)->size(), (*it)->data.c_str(), (long long unsigned int) (*it)->valid_to);
         ++it;
     }
 }
